pyproject.toml
[project]
name = "moxify-ocr"
version = "0.1.0"
description = "Custom TFLite OCR models for Moxify MTG card scanning."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "tensorflow>=2.16.0",
    "numpy>=1.26",
    "pillow>=10.0",
    "opencv-python>=4.9",
    "albumentations>=1.4",
    "requests>=2.31",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "editdistance>=0.8",
    "tensorboard>=2.16",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.4", "mypy>=1.10", "types-requests", "types-PyYAML"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/moxify_ocr"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I", "UP", "B", "C4", "SIM"]

[tool.ruff.format]
# Use ruff format as the sole formatter.

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
mypy_path = ["src", "."]
strict = true

[[tool.mypy.overrides]]
module = ["tensorflow.*", "keras.*", "cv2.*", "albumentations.*", "editdistance.*", "tqdm.*"]
ignore_missing_imports = true
